+++ inst/htmlwidgets/datatables.js
@@ -741,7 +741,38 @@ HTMLWidgets.widget({
         selected1 = $.makeArray(selected.rows);
         selected2 = $.makeArray(selected.cols);
       }
+
+      // After users reorder the rows or filter the table, we cannot use the table index
+      // directly. Instead, we need this function to find out the rows between the two clicks.
+      // If user filter the table again between the start click and the end click, the behavior
+      // would be undefined, but it should not be a problem.
+      var shiftSelRows = function (start, end) {
+    		var indexes = table.rows({ search: 'applied' }).indexes();
+    		// if start is larger than end, we need to swap
+    		if (indexes.indexOf(start) > indexes.indexOf(end)) {
+    			var tmp = end;
+    			end = start;
+    			start = tmp;
+    		}
+
+    		// i >= start && i <= end is not the right solution, because row reorder will change the
+    		// indexes order.
+    		var flag = false;
+    		return indexes.filter(function (i) {
+    			if (i === start) {
+    				flag = true;
+    			}
+    			if (i === end) {
+    				flag = false;
+    				return true;
+    			}
+    			return flag;
+    		});
+    	};
+
+
       // row, column, or cell selection
+      var lastClickedRow = 0;
       if (inArray(selTarget, ['row', 'row+column'])) {
         var selectedRows = function() {
           var rows = table.rows('.' + selClass);
@@ -756,7 +787,20 @@ HTMLWidgets.widget({
         table.on('click.dt', 'tbody tr', function() {
           var $this = $(this), thisRow = table.row(this);
           if (selMode === 'multiple') {
-            $this.toggleClass(selClass);
+            if (window.event.shiftKey) {
+          		var rows = shiftSelRows(lastClickedRow, thisRow.index());
+          		// select or de-select depends on the last clicked row's status
+          		var flagSel = !$this.hasClass(selClass);
+          		rows.each(function(value, index) {
+          		  var row = table.row(value).nodes().to$();
+          		  var flagRowSel = row.hasClass(selClass);
+          		  if ((flagSel && !flagRowSel) || (!flagSel && flagRowSel)) {
+          		    row.toggleClass(selClass);
+          		  }
+          		});
+            } else {
+              $this.toggleClass(selClass);
+            }
           } else {
             if ($this.hasClass(selClass)) {
               $this.removeClass(selClass);
@@ -773,6 +817,7 @@ HTMLWidgets.widget({
           changeInput('rows_selected', selectedRows());
           changeInput('row_last_clicked', server ?
                       DT_rows_current[thisRow.index()] : thisRow.index() + 1);
+          lastClickedRow = thisRow.index();
         });
         changeInput('rows_selected', selected1);
         var selectRows = function() {