+++ NEWS.md
@@ -10,6 +10,8 @@
 
 - Added the `RowGroup` extension (thanks, @ecortens, #466 #442).
 
+- Support using `shift + click` to select multiple rows (@shrektan, #475 #305).
+
 ## MAJOR CHANGES
 
 - For client-side tables in Shiny, the indices of selected rows (`input$tableId_rows_selected`) will include the rows that are not visible in the current view after filtering is applied. If you want to exclude the indices of such rows, you can take the intersection of `input$tableId_rows_selected` and `input$tableId_rows_all` (the latter is the indices of all rows after filtering is applied).